Improve support for `@use`
`@use` statements can specify a namespace (`@use "foo" as bar`), and the
namespace itself can be either an identifier or `*`. It can also specify a
configuration block, which is a set of parameters whose format is similar to
those of a function or mixin, but with required “defaults,” thus making them not
defaults.

Both of these concepts don't quite seem to match up to existing kinds of nodes.
`use_namespace` is an identifier, except that `*` is a valid value.
`use_parameter` is a parameter whose “default” value is mandatory. I hope that
explains why I've gone with new nodes here, and why `use_parameter` discards the
`default_value` wrapper.