Add support for maps
All reasonable usages of maps seem to work: unquoted keys, quoted keys, and even
numeric keys (which are used as an example in the docs). A key can be of any
type, so that's what's allowed in the grammar, even if the corpus doesn't test
every single possible permutation.

To pull this off, I had to exclude `:` from appearing in a `plain_value`. That's
not a problem because there aren't any bare strings for which `:` is valid…
except for an unquoted URL in a `url` function, since that's something you can
do in CSS. So bare colons are quarantined — only permitted inside of a function
called `url`.