Allow interpolation in _lots_ of places
SCSS is quite liberal in where it allows interpolation. Since most text tokens use `$._identifier` in some way, if we change all those usages to allow a mix of `$._identifier` and `$.interpolation`, that gets us most of the way there.

Interpolation inside of strings must be handled differently because of how many different kinds of characters are allowed in strings. I wrote externals to tell string segments from interpolations inside of string contexts.

There are probably some oddballs that are still unsupported simply because they're unenvisioned. I hope nobody asks for them because that would mean that they actually have a _use case_ for them, and that would confuse me.